fix: Address P0/P1 issues from comprehensive code review
P0 fixes (Critical):
- Executive task queue thread safety: Added task_mutex to protect
  shared state in executive_add_task, executive_switch_task,
  executive_complete_task
- STDP atomic weight clamping: Single fmaxf/fminf operation prevents
  torn reads during concurrent weight updates
- Corpus callosum double-free: NULL after free pattern
- Corpus callosum integer overflow: uint64_t cast before multiplication
- Amygdala numerical stability: dt_ms validation and NaN checks
- Heterosynaptic destroy: NULL after free for all resources

P1 fixes (High Priority):
- Attention memory pool thread safety: Atomic compare-exchange for
  lazy initialization prevents race condition
- Triplet STDP subnormal trace clamping: Flush traces <1e-10 to zero
  to prevent denormal arithmetic slowdown (10-100x performance hit)
- Feature extractor memory pool fallback: Heap allocation fallback
  when pool exhausted prevents unnecessary failures
- Eligibility trace NaN/Inf validation in all weight update paths
- Speech immune bridge bounds checking for array access
- Brain immune time calculation overflow prevention
- LNN layer/neuron gradient NaN/Inf checks with conservative epsilon

Test results: All 198+ tests pass across modified modules

+++ src/cognitive/immune/nimcp_brain_immune.c
@@ -130,7 +130,8 @@ const char* brain_immune_inflammation_to_string(brain_inflammation_level_t level
 static uint64_t get_timestamp_ms(void) {
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC, &ts);
-    return (uint64_t)(ts.tv_sec * 1000 + ts.tv_nsec / 1000000);
+    /* P1 fix: Cast to uint64_t before multiplication to prevent overflow on 32-bit time_t */
+    return (uint64_t)ts.tv_sec * 1000 + (uint64_t)ts.tv_nsec / 1000000;
 }
 
 /**
